+++ lib/ui/molecules/create_room/member_list.dart
@@ -1,0 +1,37 @@
+import 'package:chat_flutter/config/app_space.dart';
+import 'package:chat_flutter/ui/pages/create_room/create_room_controller.dart';
+import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
+
+class MemberList extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    final userList = Provider.of<CreateRoomController>(context).members;
+    return SizedBox(
+      height: 60,
+      child: ListView.builder(
+        scrollDirection: Axis.horizontal,
+        itemCount: userList.length,
+        itemBuilder: (BuildContext context, int index) {
+          return Container(
+            height: 40,
+            margin: EdgeInsets.all(AppSpace.small),
+            child: FlatButton(
+              onPressed: () {},
+              shape: RoundedRectangleBorder(
+                borderRadius: BorderRadius.circular(5),
+              ),
+              color: Colors.blue,
+              child: Text(
+                userList[index].name,
+                style: TextStyle(
+                  color: Colors.white,
+                ),
+              ),
+            ),
+          );
+        },
+      ),
+    );
+  }
+}
